Partial refactoring and work on Scout2Listall.py
